fixed #51

+++ engine/src/main/java/com/github/otymko/jos/runtime/VariableReference.java
@@ -16,6 +16,7 @@ import com.github.otymko.jos.runtime.machine.info.ContextInfo;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @EqualsAndHashCode(callSuper = true)
@@ -123,7 +124,7 @@ public class VariableReference extends ContextValue implements IVariable {
   }
 
   @Override
-  public float asNumber() {
+  public BigDecimal asNumber() {
     return getValue().asNumber();
   }
